+++ built_value/lib/built_value.dart
@@ -76,6 +76,14 @@ class BuiltValue {
   /// builder.
   final bool nestedBuilders;
 
+  /// Whether to auto create nested builders. Defaults to `true`.
+  ///
+  /// When this is enabled, accessing a nested builder via a getter causes it
+  /// to be instantiated if it's `null`. In most cases this is convenient, but
+  /// if you are using builders for data processing then you might need to
+  /// check for `null`. If so you should set this to `false`.
+  final bool autoCreateNestedBuilders;
+
   /// Whether builders should implement `operator==` and `hashCode`, making
   /// them comparable.
   ///
@@ -90,6 +98,7 @@ class BuiltValue {
   const BuiltValue(
       {this.instantiable = true,
       this.nestedBuilders = true,
+      this.autoCreateNestedBuilders = true,
       this.comparableBuilders = false,
       this.wireName});
 }
@@ -122,7 +131,18 @@ class BuiltValueField {
       {this.compare = true, this.serialize = true, this.wireName});
 }
 
-/// Memoized annotation for Built Value getters and methods.
+/// Optionally, annotate a Built Value `Serializer` getters with this to
+/// specify settings. This is only needed for advanced use.
+class BuiltValueSerializer {
+  /// Set this to `true` to stop Built Value from generating a serializer for
+  /// you. The getter may then return any compatible `Serializer`. Defaults
+  /// to `false`.
+  final bool custom;
+
+  const BuiltValueSerializer({this.custom = false});
+}
+
+/// Memoized annotation for Built Value getters.
 ///
 /// Getters marked with this annotation are memoized: the result is calculated
 /// once on first access and stored in the instance.